v0.12.7.2. Change codes to finish all test. Now all models have constant size. Fix all compilation warnings. Change most size_t to unsigned int to avoid compilation warnings. Update RapidJSON. Next work is to update CSparse to CXSparse.

+++ code/header/basic/area.h
@@ -11,15 +11,15 @@ class AREA : public BASE
         AREA();
         virtual ~AREA();
 
-        void set_area_number(size_t area_number);
+        void set_area_number(unsigned int area_number);
         void set_area_name(string area_name);
-        void set_area_swing_bus(size_t bus);
+        void set_area_swing_bus(unsigned int bus);
         void set_expected_power_leaving_area_in_MW(double P);
         void set_area_power_mismatch_tolerance_in_MW(double P);
 
-        size_t get_area_number() const;
+        unsigned int get_area_number() const;
         string get_area_name() const;
-        size_t get_area_swing_bus() const;
+        unsigned int get_area_swing_bus() const;
         double get_expected_power_leaving_area_in_MW() const;
         double get_area_power_mismatch_tolerance_in_MW() const;
 
@@ -32,10 +32,10 @@ class AREA : public BASE
         //double get_actual_power_leaving_area_in_MW() const;
     private:
         void set_area_swing_bus_with_zero_input();
-        void set_area_swing_bus_with_existing_bus(size_t bus);
-        size_t area_number;
+        void set_area_swing_bus_with_existing_bus(unsigned int bus);
+        unsigned int area_number;
         string area_name;
-        size_t area_swing_bus;
+        unsigned int area_swing_bus;
         double expected_power_leaving_area_in_MW;
         double area_power_mismatch_tolerance_in_MW;
 };
