+++ code/header/basic/ownership.h
@@ -12,18 +12,18 @@ class OWNERSHIP
         virtual ~OWNERSHIP();
         OWNERSHIP(const OWNERSHIP& ownership);
 
-        void append_owner_and_its_fraction(size_t owner, double fraction);
-        void delete_owner_and_its_fraction(size_t owner);
+        void append_owner_and_its_fraction(unsigned int owner, double fraction);
+        void delete_owner_and_its_fraction(unsigned int owner);
         void normalize();
 
         bool empty() const;
-        size_t get_owner_count() const;
+        unsigned int get_owner_count() const;
 
-        vector<size_t>    get_all_owners() const;
-        size_t get_owner_of_index(const size_t index) const;
+        vector<unsigned int>    get_all_owners() const;
+        unsigned int get_owner_of_index(const unsigned int index) const;
         vector<double> get_all_fraction() const;
-        double get_fraction_of_owner(const size_t owner) const;
-        double get_fraction_of_owner_of_index(const size_t index) const;
+        double get_fraction_of_owner(const unsigned int owner) const;
+        double get_fraction_of_owner_of_index(const unsigned int index) const;
 
         void clear();
 
@@ -32,8 +32,8 @@ class OWNERSHIP
     private:
         void copy_from_const_ownership(const OWNERSHIP& ownership);
 
-        bool has_owner(const size_t owner) const;
-        map<size_t,double> ownership_pair;
+        bool has_owner(const unsigned int owner) const;
+        map<unsigned int,double> ownership_pair;
 };
 
 #endif // OWNERSHIP_H
