+++ code/header/device/equivalent_device.h
@@ -14,7 +14,7 @@ class EQUIVALENT_DEVICE : public DEVICE
         EQUIVALENT_DEVICE();
         virtual ~EQUIVALENT_DEVICE();
 
-        void set_equivalent_device_bus(size_t bus);
+        void set_equivalent_device_bus(unsigned int bus);
         void set_identifier(string identifier);
         void set_status(bool status);
 
@@ -29,7 +29,7 @@ class EQUIVALENT_DEVICE : public DEVICE
         void set_equivalent_nominal_constant_power_load_in_MVA(complex<double> s);
 
 
-        size_t get_equivalent_device_bus() const;
+        unsigned int get_equivalent_device_bus() const;
         string get_identifier() const;
         bool get_status() const;
 
@@ -63,9 +63,9 @@ class EQUIVALENT_DEVICE : public DEVICE
         virtual bool is_valid() const;
         virtual void check();
         virtual void clear();
-        virtual bool is_connected_to_bus(size_t bus) const;
-        virtual bool is_in_area(size_t area) const;
-        virtual bool is_in_zone(size_t zone) const;
+        virtual bool is_connected_to_bus(unsigned int bus) const;
+        virtual bool is_in_area(unsigned int area) const;
+        virtual bool is_in_zone(unsigned int zone) const;
         virtual void report() const;
         virtual void save() const;
 
@@ -85,7 +85,7 @@ class EQUIVALENT_DEVICE : public DEVICE
         virtual void run(DYNAMIC_MODE mode);
     private:
     private:
-        size_t bus;
+        unsigned int bus;
         string identifier;
         bool status;
