+++ code/header/device/fixed_shunt.h
@@ -14,12 +14,12 @@ class FIXED_SHUNT : public DEVICE
         FIXED_SHUNT();
         virtual ~FIXED_SHUNT();
 
-        void set_shunt_bus(size_t bus);
+        void set_shunt_bus(unsigned int bus);
         void set_identifier(string identifier);
         void set_status(bool status);
         void set_nominal_impedance_shunt_in_MVA(complex<double> s);
 
-        size_t get_shunt_bus() const;
+        unsigned int get_shunt_bus() const;
         BUS* get_bus_pointer() const;
         string get_identifier() const;
         bool get_status() const;
@@ -30,9 +30,9 @@ class FIXED_SHUNT : public DEVICE
         virtual bool is_valid() const;
         virtual void check();
         virtual void clear();
-        virtual bool is_connected_to_bus(size_t bus) const;
-        virtual bool is_in_area(size_t area) const;
-        virtual bool is_in_zone(size_t zone) const;
+        virtual bool is_connected_to_bus(unsigned int bus) const;
+        virtual bool is_in_area(unsigned int area) const;
+        virtual bool is_in_zone(unsigned int zone) const;
         virtual void report() const;
         virtual void save() const;
         virtual void set_model(const MODEL* model);
@@ -45,7 +45,7 @@ class FIXED_SHUNT : public DEVICE
 
         complex<double> get_actual_impedance_shunt_in_MVA() const;
     private:
-        size_t bus;
+        unsigned int bus;
         BUS* busptr;
         string identifier;
         bool status;
